+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added project icon
 - Added metadata to Windows executable, including icon
+- Add option to check upload status
 
 ### Changed
 - **BREAKING** Dropped support for Node versions below v20